projects/hotel-dbt/requirements.txt
dbt-duckdb>=1.7.0
pandas>=2.0.0
pyarrow>=14.0.0
faker>=20.0.0
